% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{get.assay}
\alias{get.assay}
\title{Get a Project information from PRIDE}
\usage{
get.assay(assay.accession = "22134")
}
\description{
Get a project information based on pride ID from PRIDE database
}
\author{
RINTU KUTUM
}
