% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get.pride.projects}
\alias{get.pride.projects}
\title{Get all Projects}
\usage{
get.pride.projects(query = NULL, resultsPerPage = 100,
  order = "desc", speciesFilter = NULL, ptmsFilter = NULL,
  tissueFilter = NULL, diseaseFilter = NULL, titleFilter = NULL,
  instrumentFilter = NULL, experimentTypeFilter = NULL,
  quantificationFilter = NULL, projectTagFilter = NULL)
}
\description{
Get all projects lised in PRIDE database
}
