% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{get.project.assay}
\alias{get.project.assay}
\title{Get a Project information from PRIDE}
\usage{
get.project.assay(accession = "PXD000001")
}
\description{
Get a project information based on pride ID from PRIDE database
}
\author{
RINTU KUTUM
}
