% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get.project.count}
\alias{get.project.count}
\title{Get number of projects in PRIDE database}
\usage{
get.project.count(query = NULL, speciesFilter = NULL,
  ptmsFilter = NULL, tissueFilter = NULL, diseaseFilter = NULL,
  titleFilter = NULL, instrumentFilter = NULL,
  experimentTypeFilter = NULL, quantificationFilter = NULL,
  projectTagFilter = NULL)
}
\description{
Get number of projects available in PRIDE
}
\examples{
nPrideProjects <- get.project.count()
}
