% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get.project.detail}
\alias{get.project.detail}
\title{Get a Project information from PRIDE}
\usage{
get.project.detail(accession)
}
\description{
Get a project information based on pride ID from PRIDE database
}
\examples{
pID = "PXD000561"
pp <- get.project.detail(pID)
}
\author{
RINTU KUTUM
}
