% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrideProject.R
\name{getProjectDetail}
\alias{getProjectDetail}
\title{Get a Project information from PRIDE}
\usage{
getProjectDetail(projectAccession)
}
\description{
Get a project information based on pride ID from PRIDE database
}
\examples{
pID = "PXD000561"
pp <- getProjectDetail(pID)
}
\author{
RINTU KUTUM
}
