% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prideMiniList2DF}
\alias{prideMiniList2DF}
\title{List to data.frame of all PRIDE projects}
\usage{
prideMiniList2DF(x)
}
\description{
getAllPrideProject helper function, converts list to data.frame format
}
